requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-vision==2.6.3
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
multidict==5.2.0
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
yarl==1.7.2
